+++ finish/src/main/java/io/openliberty/guides/inventory/InventoryManager.java
@@ -12,9 +12,13 @@
 // end::copyright[]
 package io.openliberty.guides.inventory;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 import java.util.Properties;
 import io.openliberty.guides.inventory.client.SystemClient;
 import io.openliberty.guides.inventory.model.InventoryList;
+import io.openliberty.guides.inventory.model.SystemModel;
 import javax.enterprise.context.ApplicationScoped;
 
 // tag::ApplicationScoped[]
@@ -22,21 +26,26 @@ import javax.enterprise.context.ApplicationScoped;
 // end::ApplicationScoped[]
 public class InventoryManager {
 
-  private InventoryList invList = new InventoryList();
+  private List<SystemModel> systems = Collections.synchronizedList(new ArrayList<>());
   private SystemClient systemClient = new SystemClient();
 
   public Properties get(String hostname) {
     systemClient.init(hostname);
+    return systemClient.getProperties();
+  }
 
-    Properties properties = systemClient.getProperties();
-    if (properties != null) {
-        invList.addToInventoryList(hostname, properties);
-      }
-    return properties;
+  public void add(String hostname, Properties systemProps) {
+    Properties props = new Properties();
+    props.setProperty("os.name", systemProps.getProperty("os.name"));
+    props.setProperty("user.name", systemProps.getProperty("user.name"));
 
+    SystemModel system = new SystemModel(hostname, props);
+    if (!systems.contains(system)) {
+      systems.add(system);
+    }
   }
 
   public InventoryList list() {
-    return invList;
+    return new InventoryList(systems);
   }
 }
